# [v8.0] Optimise ASN1 decoding in X509Certificate